Multiple formula unit generation (NaCl, Na2Cl2, Na3Cl3, Na4Cl4) Energy above convex hull calculations using the Materials Project phase diagram PyMatgen space group & crystal system analysis replacing P1 defaults MACE integration for formation energies and structure relaxation Better tool transparency, now you dont just see the final report, you can also see a preview of whats being read from the json files

+++ dev/oldmcpservers/mace-mcp-server/src/mace_mcp/tools.py
@@ -530,7 +530,7 @@ def relax_structure_monitored(
             model_type=model_type, 
             size=size, 
             device=device,
-            default_dtype="float64"  # Higher precision for optimisation
+            default_dtype="float32"  # Consistent with input data
         )
         atoms.calc = calc
         
@@ -710,7 +710,7 @@ def relax_structure(
             model_type=model_type, 
             size=size, 
             device=device,
-            default_dtype="float64"  # Higher precision for optimisation
+            default_dtype="float32"  # Consistent with input data
         )
         atoms.calc = calc
